Cleaned up span converter
